refactor: Extract functions and organize code
