+++ .gitignore
@@ -12,6 +12,9 @@ dist
 dist-ssr
 *.local
 
+package.json
+package-lock.json
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
