+++ README.md
@@ -97,24 +97,25 @@ src/
 
 # 3. 개발 체크 리스트
 
-[ ] 주석이 필요 없을만큼 명확한 네이밍<br>
-[ ] 하나의 함수는 하나의 역할만 수행하기<br>
-[ ] 들여쓰기 depth는 2까지<br>
-[ ] 3항 연산자 쓰지 않기<br>
-[ ] Random 값 추출은 Random.pickNumberInRange()를 활용<br>
-[ ] 사용자의 값을 입력 및 출력하려면 Console.readLineAsync()와 Console.print()를 활용<br>
-[ ] 줄임말 쓰지 말기<br>
+✅ 주석이 필요 없을만큼 명확한 네이밍<br>
+✅ 하나의 함수는 하나의 역할만 수행하기<br>
+✅ 들여쓰기 depth는 2까지<br>
+✅ 3항 연산자 쓰지 않기<br>
+✅ Random 값 추출은 Random.pickNumberInRange()를 활용<br>
+✅ 사용자의 값을 입력 및 출력하려면 Console.readLineAsync()와 Console.print()를 활용<br>
+✅ 줄임말 쓰지 말기<br>
 [ ] 코드 포맷팅 하기<br>
-[ ] 오류를 찾을 때 함수보단 디버거 활용<br>
-[ ] 최대한 자바스크립트 API 잘 활용하기<br>
-[ ] 배열도 배열 자체 재할당이 아닌 경우 const 선언<br>
-[ ] 커밋 메세지 영어 작성<br>
-[ ] run()에서는 최대한 호출만<br>
-[ ] 비동기가 필요한 곳에만 잘 쓰였는지<br>
-[ ] Jest를 이용하여 정리한 기능 목록이 정상적으로 작동하는지 테스트 코드로 확인<br>
+[ ] 코드 컨벤션 잘 지키기<br>
+✅ 오류를 찾을 때 함수보단 디버거 활용<br>
+✅ 최대한 자바스크립트 API 잘 활용하기<br>
+✅ 배열도 배열 자체 재할당이 아닌 경우 const 선언<br>
+✅ 커밋 메세지 영어로 작성해보기<br>
+✅ run()에서는 최대한 호출만<br>
+✅ 비동기가 필요한 곳에만 잘 쓰였는지<br>
+✅ Jest를 이용하여 정리한 기능 목록이 정상적으로 작동하는지 테스트 코드로 확인<br>
 [ ] 제출할 때 회고 작성하기<br>
-(지원서에 작성한 목표를 얼마나 달성하고 있다고 생각하나요? 그 이유는 무엇인가요?
-지원서에 작성한 목표를 변경해야 한다고 생각하시나요? 그렇다면 그 이유와 어떤 목표로 변경하고 싶으신가요?
+(지원서에 작성한 목표를 얼마나 달성하고 있다고 생각하나요? 그 이유는 무엇인가요?<br>
+지원서에 작성한 목표를 변경해야 한다고 생각하시나요? 그렇다면 그 이유와 어떤 목표로 변경하고 싶으신가요?<br>
 프리코스를 진행하면서 눈에 띄는 변화나 깨달은 점이 있나요?)<br>
 [ ] 제출할 때 외부링크 걸지 말기
