--- package.json
@@ -1,33 +1,0 @@
-{
-  "name": "javascript-racingcar",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "start": "node src/index.js",
-    "test": "jest --detectOpenHandles"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.25.8",
-    "@babel/preset-env": "^7.25.8",
-    "babel-jest": "^29.6.0",
-    "jest": "^29.6.0"
-  },
-  "dependencies": {
-    "@woowacourse/mission-utils": "^2.2.0"
-  },
-  "jest": {
-    "transform": {
-      "\\.js$": "babel-jest"
-    }
-  },
-  "babel": {
-    "presets": [
-      "@babel/preset-env"
-    ]
-  },
-  "engines": {
-    "npm": ">=10.8.2",
-    "node": ">=20.17.0"
-  }
-}
